Move things to nextjs that belong there.
